[IMP] account_accountant: Allow custom tolerance on EPD reconciliation
In countries using extensively cash roundings, invoices having a early payment discount
can usually be paid within a cash rounding tolerance.
e.g. with a cash rounding of 0.05, and an early payment discount of 5%, for a total
amount of 868.25, a discount amount of 824.84 will be stored.
If a customer pays slightly more (eg 824.85), the early payment will not be recognized
in the reconciliation widget, resulting at best in a partial payment that will require
extra work for the accountant to write off properly for early payment discount.

By allowing to customize the comparison between open amount and total discount,
it allows to implement any kind of tolerance in custom modules without breaking
anything in the actual reconciliation mechanism.

OPW-3985470

closes odoo/enterprise#72367

X-original-commit: 0d2e98de241944252b88a81c8e084e16086008a1
Signed-off-by: Laurent Smet (las) <las@odoo.com>