[FIX] stock_barcode: pack lot in different packages
Steps to reproduce:

- Enable packages in the inventory settings
- Create a storable product tracked by Lot and put 4 units in stock:
    - 2 Units for "LOT001"
    - 2 Units for "LOT002"
- Create and confirm a sale order for 2 units
- Go to the barcode app > Operations > Delivery orders > your DO
- Register "LOT002" manually
> One unit will be added to your delivery moves.
- Put in pack
> You can unroll the move to see that one move line was created for 1
unit with a package and 1 unit is still missing.
- Register a second unit of "LOT002" manually to put in an other pack
> The line is updated to 2 units instead of 1 so that the additional
unit can not be packed in an other pack

Expected behavior:

You should be able to pack units of the same lot in different packages.

Cause of the issue:

To determine which line is going to be updated by our lot registration
the `_findLine` method is called. This method look for each line in
pageLines and when it finds a candidate move line it continues to find a
better candidate. However, fully packed moves should not be considered
as valid candidates.

opw-4068334

closes odoo/enterprise#68814

X-original-commit: e4c688161afb8572beea3d3c5cec1183567925cf
Signed-off-by: Steve Van Essche <svs@odoo.com>