[FIX] web_studio: fix kanban editor renderer
This commit fixes some missing features of the kanban editor renderer, that
prevents the ability to create new columns, or load more records. This was
already present in the legacy editor, but forgotten when moving to the newest
version of the editor.

QUnit tests that were testing the feature have been adapted to HOOT, while testing
those features with the newest editor instead.

Another test has been added to verify that 'load more' and 'add column' buttons
are missing as expected from Studio.

Part-of: odoo/enterprise#68084
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>