[FIX] account_accountant: use active partner in automatic reco if possible
Context: When doing automatic reconciliation, statement lines without partner but with
account_number set are using the method `_retrieve_partner` to find a partner which would
have a bank account with the same account_number.

Problem: The res.partner.bank could be active and the partner inactive, which would result
in selecting an inactive partner for the reconciliation.

This commit improve on that behavior assigning a partner if that's only one matching the bank account,
if we have more than one but only one of them is active, that's our match. The reconciliation processes
regardless of the partner being active or not since that's the only match anyway.

closes odoo/enterprise#68545

Task: 3975338
Signed-off-by: Quentin De Paoli (qdp) <qdp@odoo.com>