[REF] website: webiste controller page record is slave to listing
Before this commit, a controller page was either a listing or able to display one record.

With this commit, a controller page is now the master object that references both listing
and single record views.

That way, the route's controller has a much easy work to do and more importantly more deterministic
The dashboard to manage those website.page.controller has been replaced with a kanban view, some improvements
have been made to the form view as well.

This commit drops the support of Website Forms from the Studio interface, but making a Website Form on
an arbitrary model is reimplemented in the next commit.

task-3997658

Part-of: odoo/enterprise#68611
Related: odoo/odoo#177234
Related: odoo/upgrade#6397
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>