[IMP] mrp_mps: new yearly period + column field per date type
Currently, there are 2 fields used for how periods are displayed in the MPS:
- `manufacturing_period`: selection field with 3 options (month, week, day)
- `manufacturing_period_to_display`: integer field for how many columns are
shown for the selected period type

In prevision to the date switcher that we want to add, similar to the one
in Calendar, we're adding a new selection: year, and making one integer
field per selection.

The reason is that when switching between period types on the MPS main view,
we don't want to see the same number of columns for weeks as for years.

task 4115732

Part-of: odoo/enterprise#68548
Related: odoo/upgrade#6406
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>