[FIX] delivery_sendcloud: Enable retun label on delivery
Before this commit:

Generating return label on delivery was disabled intentionally.

After this commit:

Option to select return label on delivery is enabled.
A warning is added in `sendcloud_product_selection_widget` for customers befroe selecting return product.

According to the client related to the ticket this feature is a requierment for some marketplace users e.g Zalando, and Otto.

opw-4008265

closes odoo/enterprise#68447

Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>