[FIX] stock_barcode: dont split lines with 0 quantity
**Current behavior:**
In Barcode, it is possible for a split of incomplete moves to be
triggered which leaves the original move with a quantity and
demand of zero- effectively generating a superfluous record
value.

**Expected behavior:**
A split should not occur when the original line has
`quantity == 0`.

**Steps to reproduce:**
1. Create a transfer for 2 units of some product, assign it

2. Open the transfer in Barcode and use the form to add 1 unit

3. Use the back button within the Barcode app to return to the
     previous action

4. Reopen the transfer, set the quantity on the move to 0 from 1

5. Use the back button to exit the transfer again

6. Open the transfer in the backend to see there is a move with
     a line for 0 / 0 units.

**Cause of the issue:**
We split moves if their quantity is less than demand without
considering it may be zero.

**Fix:**
Reset the move in the case of `quantity == 0`- thus it will no
longer become a split candidate.

opw-4056241

closes odoo/enterprise#68554

Signed-off-by: Steve Van Essche <svs@odoo.com>