[FIX] account_reports: transaction without statement
Before this commit, the section transaction without statement didn't count the
reconciled entries, and so the balance of the account was wrong.
To fix that we putted a new custom engine function that will trigger the
_bank_reconciliation_report_custom_engine_common with False for the
"from_statement" and False for the "unreconciled" variables.

With that the query will get all the transaction reconciled or not that are not
from a statement.

There was also a problem if the reconcile rate was 0, which didn't when having
only unreconciled entries but now that we can have that, it needed to be
modified.

closes odoo/enterprise#68594

Task: 4110491
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>