[IMP] web_studio: improve consistency of grouped kanban
This commit brings visual consistency when toggling Studio in a grouped
kanban. Now, the column title is shown above the group, instead of showing
only the cards.

A fake progressbar is displayed if the view contains once. This one cannot
be edited yet by the editor, it is only a visual representation of the
element.

This improves the fidelity of the UI when comparing the view inside/outside
of the studio editor.

Existing tests have been modified to assert new elements are present in
the view.

closes odoo/enterprise#68084

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>