[IMP] account_accountant: bank rec UI improvement
Before, in the bank reconciliation widget, when clicking on the
statement total, the form view of the statement was opened in a dialog.
This doesn't look good anymore after 120a546295b9e99b2bd50736003506311419562e as that
commit adds the chatter and the attachment previewer to the form view.
This commit removes the chatter and the attachment previewer from the
dialog.

No task.

closes odoo/enterprise#68563

Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>