[FIX] voip: adapt voip model with model fix
The voip model was overriding `static insert` method as if this
works on a single record. For some time, this method works on
multiple records, and its implementation detail has changed so
assuming it does `get() ?? new()` is no longer guaranteed.

The intent of override was to enrich data after them being assigned.
This has been converted to an override of `update()`, which
guarantees it being called whenever fields are updated on record,
without making too much assumption in implementation details of
records.

This commit also speeds up tests of VOIP, which were awaiting
input that contains a value. This is not observed by mutation
observer, so these tests took 3 seconds to execute. This PR puts the
value in `data-value`, so that this is a mutation that can be
observed by contains, thus reducting time of such test to mere dozens
of ms.

closes odoo/enterprise#68604

Related: odoo/odoo#177059
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>