[FIX] hr_contract_salary: don't delete employee having contract
Once an archived employee has a draft contract archived, he is deleted.

We should not do that as it tries to delete some of employees that
really have worked in the company.

So we aslo check that the employee we want to delete has no other
contract in the company.

We also don't need to look after cars, as they are not created if the
contract is not signed.

closes odoo/enterprise#69088

X-original-commit: 7941cac30a2397957d92d95da19b83733e57defa
Signed-off-by: Pierre Masereel (pim) <pim@odoo.com>