[FIX] industry_fsm_stock: Fix singleton error
Steps to reproduce:
In field service, have a project on which you have to sell 3 of item A.
Item a is tracked by serial number and the delivery route is in 2 steps.
From stock to output and from output to customer.

On the sale order make sure you have 3 separate lines of 1 quantity of
product A separated or not by sections.
Confirm this one. This will create two pickings.
One picking linked to a global stock move for 3 units from stock to output.
A second picking linked to 3 stock moves each linked to a sale line id.
Prioritise the first one by starring it.

Now go back to the project and mark is as done.

OPW-4088696

Had to redo the tests as from 17.2 one picking is generated at a time.
Here we modify rules to create the two pickings.
We have to do this to mimic customers migrating from prior versions
of 17.2 to later versions.

closes odoo/enterprise#69222

X-original-commit: e49e4805dbd099ed9dd4c579518a70eaec655813
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>