[IMP] l10n_ec_edi : enhance error message for certificates
Steps to reproduce:
- Install the module "l10n_ec" and switch to ecuadorian company
- Go to Accounting > Configuration > Ecuadorian SRI > Digital Certificates SRI
- When uploading a p12 file with the country name "Ecuador" and not "EC"
- An error pops up but does not tell the right reason

Cause:
The cryptography library only accepts country names in the form of 2 character country code so it raises an error.
If an error occurs during the loading of the file, the error message will always be "Error loading certificate %s, check that password is correct and file type is p12".

Solution:
Display the error returned by the cryptography library. If the error is the password it will display "Invalid password or PKCS12 data" so I removed the last part of the old message. Now when the error comes from the country name it will display: "Country name must be a 2 character country code"

opw-4116914

closes odoo/enterprise#69138

X-original-commit: b232271fdcf5752a5d66062f72d52a66610a4b3a
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Mathieu Coutant (mcou) <mcou@odoo.com>