[IMP] account_reports,l10n_{uk,ae,nz}_reports: Add tax closing start date
Some countries need to support another start date than first January.
This commits aims to provide that by adding a company_dependent field on the report that each company could configure.
A new date filter `tax_period` appeared with this feature. So now, the normal way to address periods on tax reports is by using this filter.
If we try to close a period without this a warning appears telling us we are not using the filter and we can still proceed or go back.

To compute the period boundaries two functions need to stay consistent with each other, one in python `_get_tax_closing_period_boundaries`
and the other in javascript `_computeTaxPeriodDates` so that the filter is showing the same result as the report.

When companies are using a tax unit, the main company of that tax unit will serve as reference for the periodicity and the start date.

task-4110457

Part-of: odoo/enterprise#70076
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>