[FIX] account_avatax*,*external_tax*: no external tax on down payments
This partially redoes odoo/enterprise#64900. Copying tax_ids from the
sale order lines to the down payment line on the invoice doesn't work
for external tax calculators. Tax percentages returned by tax
calculators are only informational. When they return e.g. 5% tax, the
actual amount can be different from 5% because of differences in
rounding or (partial) exemptions. The authoritative tax is returned as
a value amount (e.g. $10). Because of this, we cannot rely on Odoo to
calculate the tax from account.tax records. It's not always going to
match.

It's not possible to fix this by externally calculating taxes on the
down payment invoice. The single down payment line cannot accomodate
all different tax categories. Creating a down payment line per tax
category (e.g. avatax_category_id) was considered, but a specific down
payment product would need to be created per used tax category which
would clutter the database. On top of that, down payment products are
removed in later versions [1] which would complicate this approach
even further.

This solves the issue by not calculating taxes on down payments at
all. The down payment invoice will not be sent to external tax
calculators. The resulting down payment lines on the SO won't have
taxes set, and those lines will be excluded from the request to the
external tax calculator. The final, "regular invoice" transaction will
happen as if no down payments were made. The down payment lines won't
be sent, so this will look like a regular invoice from the external
calculator's view.

By default, the downpayment lines equal the subtotal of the selected
down payment amount. We remove tax_ids from the lines because we
cannot count on calculating tax locally (see above). To keep the
behavior analogous we override the downpayment wizard to create a
single down payment line with a total equal to the entered total in
the wizard.

[1] odoo/odoo@9aa52dd6418e5881adc2d96d15d062b55d6150c5

opw-3971513

closes odoo/enterprise#73110

X-original-commit: 2bb532773650650fad01b538c0395c9796069cc3
Related: odoo/odoo#185938
Signed-off-by: Joren Van Onder (jov) <jov@odoo.com>
Signed-off-by: William André (wan) <wan@odoo.com>