[FIX] appointment_google_calendar: prevent access error during sync
Before this commit, users could encounter an access error when
attempting to connect to Google Calendar. If a user was removed from an
appointment but an existing calendar event still belonged to that user,
all attendees of that event would be unable to sync their calendar with
Google, resulting in an access error. The error message displayed was:
Sorry, Marc Demo (id=6) doesn't have 'read' access to:
- Appointment Type, Test Demo (appointment.type: 3).

opw-4109304

closes odoo/enterprise#68711

X-original-commit: e93f39e6664596f45ebb31f3a92c21fab1169cf7
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Pedram Bi Ria (pebr) <pebr@odoo.com>