[IMP] mrp_mps: add period switcher on main view
This commit adds a date switcher like the one in calendar. It allows to switch
between daily/weekly/monthly/yearly display on the MPS main view without
needing to go to the settings every time.

It does not change the value of `manufacturing_period`, which is the default
setting used when opening the MPS view.

It also ensures that the number of columns to display must always be positive.

task 4115732

closes odoo/enterprise#68548

Related: odoo/upgrade#6406
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>