[FIX] account_accountant: tour fails on clicking on the pager
The tour changes the number of elements in the pager.
We then check that the number of elements do change in the list.
But the button stay disabled longer, which implies that we try
to click on it too soon in the tour, failing it.
We should wait for the button to be enabled before clicking on it.

runbot-74835

closes odoo/enterprise#68701

Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>