[FIX] l10n_ke_edi_oscu_stock: fix traceback when creating PO
Currently, a traceback occurs when the user creates a PO
with a product from a company other than 'KE'.

To reproduce this issue:

1) Install 'l10n_ke_edi_oscu_stock'
2) Switch to a company other than 'KE'
3) Create a PO with a product > `Confirm Order`
4) Click on `Receive Products` > `Validate` and get back to PO through breadcrumb
5) Click on the `Create Bill`

Error:-
"ValueError: Compute method failed to assign purchase.order.line(19, ).display_name"

The above error is occurring because of a compute method in which
A filter is used to access the 'KE' company record.

But here the user is not in 'KE' company so it returns None.
But the 'display_name' is a readonly and nonstored filed.

Which leads to the above traceback.

https://github.com/odoo/enterprise/blob/7a9157e059fd70831eb3174863a797abdb99ad53/l10n_ke_edi_oscu_stock/models/purchase.py#L71-L73

By applying this commit, We can resolve this issue by
giving the POl  name when it is not 'KE' company.

sentry-5726618591

closes odoo/enterprise#69241

X-original-commit: e2521ebe379fa279e7cd66dc2b6fbbb9b3a72379
Signed-off-by: Josse Colpaert (jco) <jco@odoo.com>
Signed-off-by: Altaf Shaik (alsh) <alsh@odoo.com>