[IMP] stock_barcode: avoid display same notif
In the Barcode app, it is possible to scan multiple barcodes at once.
So, it can happen the same notification to be displayed multiple times
in a row. For example:
- Open a delivery where the source location is mandatory;
- Scan multiple serial numbers at once (don't scan the source before.)
The notification asking to scan the source first will be displayed for
each scanned serial numbers.

To avoid that, keep displayed notification in cache and don't display
one if it was already displayed during the processing of the current
scanned barcode.

task-4089638

Part-of: odoo/enterprise#68825
Related: odoo/odoo#177738
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>