[IMP] sale_subscription: pending transaction flag removed with post invoice
Before this commit, subscriptions automatic invoices are skipped when the flag pending transactions is set.
it is correct to avoid double payment for the same period but there is a problem if the flag stays while the subscription has been renewed since.

This commit aimes to remove that flag when an invoice is posted for this
subscruption.

closes odoo/enterprise#68713

Taskid: 4107185
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>