[IMP] project_enterprise,industry_fsm: review dates in task kanban view
Currently, the planned dates of a task takes too much space on the kanban card and could be rendered on 2 lines instead of one.
The goal of this task is to reduce a bit the spaces taken by the planned dates in the kanban card of tasks.

More specifically, we remove the seconds from the dates to gain space and we handle the following cases:
- When the begin date and end date are the same:
We shorten from this "08/23/2024 09:15:00 -> 08/23/2024 11:00:00" to this "08/23/2024 09:15 -> 11:00" (remove the end date)
- When they are not:
We shorten from this "08/23/2024 09:15:00 -> 08/30/2024 11:00:00" to this "08/23/2024 -> 08/30/2024" (remove the hours)

task-4120069

closes odoo/enterprise#68708

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>