[FIX] web_gantt: tests - fix some mobile issues
In the community counterpart of this commit, the `pointerType`
properties of pointer events have been set properly according to the
current environment.

This change means that drag & drop will now wait for a "touchDelay" in
mobile environments before the drag sequence initializes. Because of
this, some Gantt tests failed and needed adaptation.

X-original-commit: b8bfaafc8bade115dca21d42347ee22d450d719f
Part-of: odoo/enterprise#69213
Related: odoo/odoo#178625
Signed-off-by: Michaël Mattiello (mcm) <mcm@odoo.com>
Signed-off-by: Julien Mougenot (jum) <jum@odoo.com>