[FIX] pos_blackbox_be: fix insz check for people born after 2000
If you were born after 2000 your insz number wouldn't be recognized as
valid.

Steps to reproduce:
-------------------
* Install pos_blackbox_be module and HR module
* Create an employee with a birth date set after 2000
* Generate an insz number for a person born after 2000
  https://www.rsolution.be/rijksregister-nummer-generator.RSolution
* Set the number on the employee and remove all `.` and `-`
> Observation: When saving the insz wouldn't be recognized as valid

Why the fix:
------------
When trying to validate a insz number for a person born after 2000, the
calculation is a bit different. You need to add a `2` before
`partial_number` to make it work.
Explanation can be found on wikipedia here:
https://fr.wikipedia.org/wiki/Num%C3%A9ro_de_registre_national

opw-4077545

closes odoo/enterprise#68723

Signed-off-by: Vlad Stroia (vlst) <vlst@odoo.com>