[IMP] sale_subscription: prevent assignment mail for recurring invoices
Before this commit, assignment emails are sent when an account.move is
created with a different user_id than the current one. For subscriptions
it can send thousands of email depending on the database size.

The user_id on the account.move is the same salesperson than the one in
charge of the sale.order. It helps the consistency of communication
between the customer and the company.

closes odoo/enterprise#68752

Taskid: 4096474
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>