[FIX] planning: dont go over 59.
round() on the fractional part of shift_template.end_time can lead to problems when the result is very close to 1.0 (e.g., anything greater than 0.994).
Specifically, round() can cause the fractional part to round up to 1.0 (so 60 minutes), which is not wanted.
We can either set it to 59 as done here, or 0 and add 1 hour to the hours.

closes odoo/enterprise#68791

X-original-commit: 8be9595417648d08e91001365f085b6ef511b219
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>