[FIX] website_helpdesk: don't alter the subtype when insert attachment
Before this commit, when the helpdesk ticket is created by the website
form and need to insert an attachment, the subtype used is altered to be
able to let the portal user to see it instead of a using a non internal
subtype.

This commit reverts the changes made in d9b06557baf692f707ff8b75106df6b67c6bc523
to avoid altering the subtype.

closes odoo/enterprise#68792

X-original-commit: 50cc42414528cc2061a7b9c861f522458ae622c4
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>