[FIX] mrp_subcontracting_account_enterprise: correctly calculate report
Steps to reproduce the bug:

- Create a storable product P1 with the following BoM:
    - Type: Subcontracting
    - Subcontractor: Azure Interior
    - Components:
        - C1: 1 unit (price of C1: $10)
    - Cost of P1: $10

- Create a purchase order:
    - Product: P1
    - Quantity: 100 units

- Confirm the purchase order.

-  Receive 99 units and create a backorder for 1 unit.

- Go to Manufacturing > Reporting > Production Analysis.

Problem:

The "Total Subcontracting Cost / Unit" is $40 instead of $20. This
occurs because, in the SQL query, when performing the left join
between `stock_move_move_rel` and `stock_move`, the moves appear
duplicated as they have the same `move_orig_id`. Consequently, when
calculating the total quantity, it results in
`(99 + 99) = 198 * $20 = $3960`, but the total is only divided by the
quantity of one move, leading to $3960 / 99 = $40.

opw-4061947

closes odoo/enterprise#68834

X-original-commit: ffb051b354f72f97452ad03373cf1d205fccc43f
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Djamel Touati (otd) <otd@odoo.com>