[FIX] sale_subscription: get oldest valid child for reinvoice
**Current behavior:**
After creating a subscription for a recurring product which
creates a project/task, invoicing the original order and posting
it, then creating the renewal order, if you create a purchase
order for some expensable service product that has an
`expense_policy == 'sales_price'` and attempt to create the
vendor bill for the purchase order, you will get a UserError
about the original sale order being locked.

**Expected behavior:**
Because there is a renewed order which is not locked, the
expense should use that instead of the original subscription
order.

**Steps to reproduce:**
1. Create a recurring service product which is set to create a
     project and task that has an associated analytic account

2. Create a subscription for the product and confirm it, create
     the invoice and post it (locking the order)

3. Renew the subscription, confirm it, create the invoice but do
     not post it

4. Create a service product which can be expensed and uses the
     expense policy of sales price

5. Create a purchase order for the expensable service and set
     analytic distribution on the order line to use the same one
     from the subscription

6. Confirm the purchase order and create an invoice, then try to
     post it -> UserError

**Cause of the issue:**
The original subscription order is found via the shared analytic
account and generates the error because it is now in a locked
state.

**Fix:**
Check if a found order is in the renewed subscription state and
then search for the lowest id order in DB with the originally
found subscription as its origin order- if one exists, use this
record instead of the original order in the
`_sale_determine_order()` mapping.

opw-4061484

closes odoo/enterprise#68820

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>