[FIX] stock_barcode: negative qty fulfill button
Fixes minor visual issue with the fulfill button when more than expected
quantity was scanned. In such a case, the quantity to fulfill is
negative and the quantity displayed on the button was written as "+-N"
instead of simply "-N".
The issue was the plus symbol was always displayed, no matter the
quantity value.

Part-of: odoo/enterprise#68825
Related: odoo/odoo#177738
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>