[IMP] stock_barcode: set lot quant qty on theoric
Before this commit, during an Inventory Adjustment in the Barcode app,
if a barcode for tracked by lot product is scanned alonside some
quantity (a GS1 barcode for example), the barcode's quantity was set on
the quant line.

With this commit, the theorical quantity will be used instead.

The idea is, if a quant already exists, its quantity was maybe modified
since the quant's creation. But in any case, the GS1 barcode label was
surely not re-printed and so, the barcode still refers to its original
quantity. During an inventory adjustment, we don't care about the
original quant quantity, we only care about actual quantity.

task-4089638

Part-of: odoo/enterprise#68825
Related: odoo/odoo#177738
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>