[IMP] account_reports: Add the possibility to define custom labels in _get_custom_groupby_map
Groupby lines generated from _get_custom_groupby_map were displayed with a name directly identical to the value used for the grouping key. When that value was just a technical label (needing for example to be replaced by a translated string), or a json serialization of a set of values, this required the use of a _custom_line_postprocessor to recompute those lines' names, which wasn't really convenient. We hence introduce a new key in the custom groupby map, allowing specifying a function to be called on the custom grouping keys, to generate their UI label.

Part-of: odoo/enterprise#60153
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>