[FIX] account_accountant: AML view opening form instead of preview
The bug consisted of the form view to be opened when we clicked on an aml (journal item).

The expected behaviour is to preview its attachment if there is one. But we never want to show that form view

The cause is a commit from RD-JS
https://github.com/odoo/odoo/pull/176707/files

That changed the logic of opening the form view if there is no actions.

task-4132306

closes odoo/enterprise#68835

Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>