[FIX] industry_fsm_sale : Avoid modifying invoice_status after invoicng SO
Steps to reproduce:
	- Install Sale and Field Service modules
	- Create a Quotation with a normal product and another Field service product
	- Confirm the quotation and invoice it
	- Change the delivered quantity of the field service product to meet the order quantity
	- Check the invoice status (will be 'Fully Invoiced')
	- Set the field service product in the SO to a unit price 0

Current behavior before PR:
After creating an invoice for a SO that has field service product and then
set the product's unit price to 0 the SO invoice_status will change from
'Fully invoiced' to 'Nothing to invoice'. This is happening because in FSM
we don't invoice SOL that has price_unit equals 0 which will lead that the SO
inovice status will change too to 'Nothing to invoice'.
https://github.com/odoo/enterprise/blob/17.0/industry_fsm_sale/models/sale_order.py#L84:L91

Desired behavior after PR is merged:
We are now checking before computing the invoice status of any FSM product if the SO
is not yet invoiced because if it is we won't modify the invoice status of the SOL.

opw-4017282

closes odoo/enterprise#69483

X-original-commit: b58f71fdee047fa101e1ce843ed65e33076a9411
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Youssef Bashandy (yoba) <yoba@odoo.com>