[FIX] hr_work_entry_contract_enterprise: work entries generation
With https://github.com/odoo/enterprise/pull/55618, the computation of
the start and ending dates in the gantt view has slighly changed. For
example, globalStop, replacement of stopDate, is now computed as the
start of a period instead of the end of a period. In the end, globalStop
and stopDate differ by 1 millisecond which is not a problem at all
when displaying the gantt records but is a problem when generating
work entries: they are generated for an extra day.

We fix that problem by removing 1 millisecond to globalStop where it is
used in the work_entries_gantt view.

X-original-commit: cfcc99cf70f5abf5ccfa0c866103060e078ce34b
Part-of: odoo/enterprise#68971
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>
Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>