[FIX] *: gantt: no limit on records
For some views like the planning_gantt and work_entries_gantt views, it
appears to be a bad idea from a functionnal point of view to be limited
to 1000 records displayed. Here we simply drop the limit mechanism. If
we meet some problem with that in the future, we will maybe reintroduce
it in some particular views.

Note that the bound fixed by Firefox (mentioned in https://github.com/odoo/enterprise/pull/55618) on the number of rows/columns
in the css grid is actually never reached since the
number of pills to display is limited by the maximun number of pills that
can be put in a bit more that a screen (see isOut in computeVisiblePills).

closes odoo/enterprise#68971

X-original-commit: 916aa6cc8c4690bae8b3cceb7ad8e951e8ca4aee
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>
Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>