[FIX] industry_fsm: display fsm tasks with privacy invited only
Steps to reproduce:
- Log in as admin
- Project > Configuration > Projects > Field service
- In the Settings tab set Visibility to 'Invited Internal Users'
- Create a task with Marc Demo as assignee
- Check that Marc Demo is a follower of the task and not of the project
- Settings > Users & Companies > Users > Marc Demo
- Set Marc Demo's  access rights for Project and Field Service to 'User'
- Log in as Marc Demo > Field Service
- No tasks are displayed

As per the specifications of the 'Invited Internal Users' mode of project privacy, a
user should be able to see tasks they are a follower of even if they do
not have access to the project they are from. This is correctly reflected in
Projects > Tasks > My Tasks where the followed task is visible.

This does not happen because of a lack of access rights on the task as
we can copy the url of the task in Field Service as admin and access the
page when pasting it as demo.

The issue here is twofold:
- is_fsm is currently a field defined with both 'related' and 'search'
arguments. Prior to 17.2, search is ignored in favor of search_related
so the domain once again tries and fails to access the information on
the related project. The search function defined here (which was
previously unused) bypasses the issue by getting is_fsm through an SQL
query.

- Since is_fsm is not storable on task, the search function is used to
translate domains of type ('is_fsm', '=', ...). However the
_search_is_fsm function did not take into account the value passed as
argument (only the operator), resulting in domains like ('is_fsm', '=', 'False') not being
inverted as they should.

opw-4003133

closes odoo/enterprise#69071

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>