[FIX] sign: send mail after transaction commit
As we force sending the mail to the signatories, we've put a commit in
the transaction before sending the mail to ensure the document is signed
before sending the mail.

The issue is that if the transcation failed later in an override (e.g.
create a car, send another document, ...), the document is signed and
the transaciton has only been partially processed.

To avoid this kind o issues (silent errors), we are using the function
'send_after_commit' instead of 'send' to only send the email when the
entire transaction is processed.

closes odoo/enterprise#69082

X-original-commit: 9d99c3e3632be3b8136925961df285c308b0edc7
Signed-off-by: Pierre Masereel (pim) <pim@odoo.com>