[FIX] account_reports: fix generation of subheaders
Steps to reproduce:
- Install l10n_co_reports
- Switch to a Colombian company (e.g. CO Company)
- Go to "Accounting / Reporting / Colombian Statements / Certificado de Retención en IVA"

Issue:
A traceback is triggered:
"UncaughtPromiseError > OwlError
Caused by: TypeError: Cannot read properties of undefined (reading 'colspan')
..."

Cause:
When generating the subheaders from the columns, the case (where the
column is not part of a group key and its "comparison_mode" attribute
is not set) is not handled.
Therefore subheader has "Undefined" as value, which generates an
error when trying to access its "colspan" property in the template.

Solution:
When generating the subheaders, always return the object with the default
values when the column is not part of a group key.

closes odoo/enterprise#68990

Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>