[IMP] account_reports: when closing an entry tax_lock_date must not interfere with the moves
When we clicked on the closing entry, we would not get the moves that have to be generated by a locked company.
The reason was that we were generating moves only for the non locked companies and recomputing move_lines for draft moves in locked companies. So if the company was locked without an existing closing move we would not be able to create a new one.

An additional improvement come with this. The improvement is that we do not recompute tax closing entries if the move is already created.
This prevent from overriding user's input.

task-4110459

Part-of: odoo/enterprise#68551
Related: odoo/odoo#180095
Related: odoo/upgrade#6390
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>