[FIX] pos_preparation_display: add sound assets
In commit 45f568f8f8252913ed7cba00fe3da6f48edb7b26 we replaced the pos
sound service with the mail sound service.

In this commit we ensure that the correct assets are loaded in
`pos_preparation_display`. The assets that handled the specific
notifications used in pos were not loaded.

closes odoo/enterprise#70245

Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>