[FIX] sale_stock_renting: allow visibility days and renting
The case:
- I want to organise all my shop in order to be able to serve all the
  rental orders I planed for a week.

The issue, rental orders create a retun picking directly and their
forecast could easily reach zero. (even more on small renting period)

An existing solution:
Correctly set a small lead time on reordering rule in order to have a
fix day in order to have the forecast. However it only works if people
check the replenishment report everyday and don't have weekend.
And if they use automatic RR they lose the control over the warehouse
that should resupply the product.

Also they can't use the visibility days feature and plan for a week.

In order to allow a better planning and flexibility we add a small
parameter to ignore the returns of rental order. That way the forecast
ignore returns and allow visibilty.

closes odoo/enterprise#69482

X-original-commit: 47959e1636f5996494994828765c78a036d19d4b
Signed-off-by: William Henrotin (whe) <whe@odoo.com>