[FIX] social: display number of records in kanban
Before this commit, there were discrepancies between a kanban view with
or without the progress bar. The kanban view, with the progress bar,
have the count of each group (folded or not). Contrariwise, the kanban
view, without the progress bar, only have the count of the folded groups.

This commit adds a count to all the columns of the kanban view, folded
or not, with the progress bar or without.

opw-4132389

closes odoo/enterprise#69097

X-original-commit: 6e60a1f19f6bf7bdf68776048254322a8c7c8061
Related: odoo/odoo#178271
Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>