[IMP] l10n_ke_edi_oscu: Increase timeout to 45s and improve error msg
We noticed that in production, eTIMS regularly fails to respond within
the timeout of 30s when we send an invoice.

Since this happens fairly often, this PR aims to improve the error
message so that the user knows there is nothing they can do except
wait for a bit and then retry.

We also increase the timeout to 45s, on the off chance that this might
improve things somewhat.

closes odoo/enterprise#69049

Taskid: none
Signed-off-by: Josse Colpaert (jco) <jco@odoo.com>