[FIX] product_barcodelookup: fix traceback when scan a barcode without description
Currently, a traceback occurs when the user create a product with barcode
of having no description.

To reproduce this issue:

1) Install stock, barcode lookup
2) Give an API key in settings for barcodelookup
3) Now create a new product variant by scanning the barcode of a product
   or enter the barcode manually.

Note:-
API key for testing :- `wjgkytojaghoiyzu9n2f6vc6wgysp9`
Barcode for testing:- 8901207035715

Error:- "TypeError: unsupported operand type(s) for +=: 'bool' and 'Markup'"

Initially, there will be no product description while creating a new product.
So it leads to a traceback when `+=` is used between the markup and description

https://github.com/odoo/enterprise/blob/67ecb18497ee3f1ade6cc81738db3df08f94ba2f/product_barcodelookup/models/product_template.py#L189-L190

sentry-5744282512

closes odoo/enterprise#69252

X-original-commit: dbddf99b72ab5a624242429e9bcb02130c26e0f5
Signed-off-by: Adrien Widart (awt) <awt@odoo.com>
Signed-off-by: Altaf Shaik (alsh) <alsh@odoo.com>