[FIX] sale_purchase_stock_inter_company_rules: force readonly
Because of
https://github.com/odoo/odoo/blob/e1fb54cb3d5db9c906946fa11ff49c1ca86d3722/odoo/fields.py#L620-L621,
writing on a field which is `related` and `readonly=False` will
effectively forward the new value on the parent field if this one is
also `readonly=False`. This was not the case for fields

 + `intercompany_warehouse_id`
 + `intercompany_sync_delivery_receipt`
 + `intercompany_receipt_type_id`

defined in res_company and related in res_config_setting

As they are computed fields which implies that are `readonly=True` by
default.

closes odoo/enterprise#69133

X-original-commit: 9e4b3677faa8db19d07cb4e35eaad354d992b5f3
Signed-off-by: William Henrotin (whe) <whe@odoo.com>