[IMP] *: form: nolabel fields should take all width
Grid is set to use 2 columns in a `<group/>` used as an inner group.
So before this commit, we had to put `colspan="2"` to force no label
fields to take all the available space.
If not, the field was located in the first column and you had some
empty space on the right due to the second column.

In this commit, we want to make this attribute optional when
`nolabel="1"` by collapsing these 2 columns automatically.

<group string="...">
    <field name="x2many_field" nolabel="1"/>
</group>

closes odoo/enterprise#69070

Related: odoo/odoo#178198
Signed-off-by: Pierre Paridans (app) <app@odoo.com>