[IMP] website_studio: snippet option to expose model as page
This commit revamps the logic of the creation of Website Forms in studio.

Namely the creation of Forms is now handled as it should be, that is in Website itself.
This commit adds that feature on the relevant website's Snippet Option.

part of task-3997658

closes odoo/enterprise#68611

Related: odoo/odoo#177234
Related: odoo/upgrade#6397
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>