[IMP] quality_control: add quality checks with spreadsheets
This commits add a new way to design quality control checks with
spreadsheets.

It can be used similarly to worksheet but doesn't require studio and it might
be easier to configure.

The value of a cell is used as the result of a quality control point.
The check passes if the value is truthy (TRUE, number other than 0) and
fails otherwise.

This value can be computed based on the value of others cells with formulas.

It allows to implement custom logic and interactions between
different inputs (while worksheets are static).

Garbage collect history
-----------------------
Quality checks with spreadsheets can generate a lot of spreadsheets
and consequently a lot of spreadsheet revisions.

We don't particularly care about the history for quality checks,
as they are mostly one-time-use spreadsheets and only the final
result matters.

The history is cleared after one day.
Otherwise, the database could be grow and grow needlessly.

Task: 4144096
Part-of: odoo/enterprise#69141
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>